+++ src/Sentry/Internal/Constants.cs
@@ -40,6 +40,8 @@ namespace Sentry.Internal
         // See: https://github.com/getsentry/sentry-release-registry
 #if ANDROID
         public const string SdkName = "sentry.dotnet.android";
+#elif IOS || MACCATALYST
+        public const string SdkName = "sentry.dotnet.cocoa";
 #else
         public const string SdkName = "sentry.dotnet";
 #endif