+++ src/Sentry/Scope.cs
@@ -5,6 +5,7 @@ using System.Linq;
 using System.Threading;
 using Sentry.Extensibility;
 using Sentry.Internal;
+using Sentry.Internal.Extensions;
 
 namespace Sentry
 {
@@ -326,8 +327,7 @@ namespace Sentry
         public void Apply(IEventLike other)
         {
             // Not to throw on code that ignores nullability warnings.
-            // ReSharper disable once ConditionIsAlwaysTrueOrFalse
-            if (other is null)
+            if (other.IsNull())
             {
                 return;
             }
@@ -390,8 +390,7 @@ namespace Sentry
         public void Apply(Scope other)
         {
             // Not to throw on code that ignores nullability warnings.
-            // ReSharper disable once ConditionIsAlwaysTrueOrFalse
-            if (other is null)
+            if (other.IsNull())
             {
                 return;
             }